feat(API): added All needed APIs for works,posts,products

+++ app/Http/Controllers/ProductController.php
@@ -9,6 +9,9 @@ use App\Models\Product;
 use Illuminate\Http\Request;
 use Illuminate\Routing\Controllers\HasMiddleware;
 use Illuminate\Routing\Controllers\Middleware;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Storage;
+use Illuminate\Validation\Rule;
 
 class ProductController extends Controller implements HasMiddleware
 {
@@ -17,7 +20,7 @@ class ProductController extends Controller implements HasMiddleware
      */
     public function index()
     {
-        return ProductAllResource::collection(Product::paginate(12));
+        return ProductAllResource::collection(Product::orderByDesc('id')->paginate(12));
     }
 
     /**
@@ -25,7 +28,34 @@ class ProductController extends Controller implements HasMiddleware
      */
     public function store(Request $request)
     {
-        //
+        $request->validate([
+            'image' => ['required', 'image', 'mimes:jpeg,png,jpg,gif', 'max:2048'],
+            'title' => ['required', 'string', 'min:3'],
+            'slug' => ['required', 'string', 'min:3', 'unique:products'],
+            'price' => ['required', 'integer', 'min:0'],
+            'price_with_discount' => ['nullable', 'integer', 'min:0', 'lt:price'],
+            'expert' => ['required', 'string'],
+            'description' => ['required', 'string'],
+            'download_link' => ['nullable', 'string'],
+        ]);
+
+        $user_id = Auth::guard('sanctum')->user()->id;
+        $product = Product::create([
+            'user_id' => $user_id,
+            'image' => $request->file('image')->store('products'),
+            'title' => $request->title,
+            'slug' => $request->slug,
+            'price' => $request->price,
+            'price_with_discount' => $request->price_with_discount,
+            'expert' => $request->expert,
+            'description' => $request->description,
+            'download_link' => $request->download_link,
+        ]);
+
+        return response()->json([
+            'message' => __('messages.productCreated'),
+            'product_id' => $product->id
+        ], 201);
     }
 
     /**
@@ -49,7 +79,38 @@ class ProductController extends Controller implements HasMiddleware
      */
     public function update(Request $request, Product $product)
     {
-        //
+        $request->validate([
+            'image' => ['sometimes', 'image', 'mimes:jpeg,png,jpg,gif', 'max:2048'],
+            'title' => ['required', 'string', 'min:3'],
+            'slug' => ['required', 'string', 'min:3', Rule::unique('products')->ignore($product->id),],
+            'price' => ['required', 'integer', 'min:0'],
+            'price_with_discount' => ['nullable', 'integer', 'min:0', 'lt:price'],
+            'expert' => ['required', 'string'],
+            'description' => ['required', 'string'],
+            'download_link' => ['nullable', 'string'],
+        ]);
+
+        $oldImagePath = $product->image;
+
+        $product->update([
+            'image' => $request->hasFile('image') ? $request->file('image')->store('products') : $product->image,
+            'title' => $request->title,
+            'slug' => $request->slug,
+            'price' => $request->price,
+            'price_with_discount' => $request->price_with_discount,
+            'expert' => $request->expert,
+            'description' => $request->description,
+            'download_link' => $request->download_link,
+        ]);
+
+        if ($request->hasFile('image') && $oldImagePath) {
+            Storage::delete($oldImagePath);
+        }
+
+        return response()->json([
+            'message' => __('messages.productUpdated'),
+            'product' => $product
+        ], 200);
     }
 
     /**
@@ -57,8 +118,14 @@ class ProductController extends Controller implements HasMiddleware
      */
     public function destroy(Product $product)
     {
+        if ($product->orders()->count() > 0) {
+            return response()->json([
+                'message' => __('messages.cannotDeleteProductWithOrder'),
+            ], 400);
+        }
+
         $product->delete();
-        return response()->json(null, 204);
+        return response()->json(['message' => __('messages.productDeleted')], 200);
     }
     public static function middleware(): array
     {
